+++ components/terms_of_service/terms_of_service.test.tsx
@@ -31,62 +31,62 @@ describe('components/terms_of_service/TermsOfService', () => {
 
     test('should match snapshot', () => {
         const props = {...baseProps};
-        const wrapper = shallow(<TermsOfService {...props}/>);
+        const wrapper = shallow<TermsOfService>(<TermsOfService {...props}/>);
         expect(wrapper).toMatchSnapshot();
     });
 
     test('should call getTermsOfService on mount', () => {
         const props = {...baseProps};
-        shallow(<TermsOfService {...props}/>);
+        shallow<TermsOfService>(<TermsOfService {...props}/>);
         expect(props.actions.getTermsOfService).toHaveBeenCalledTimes(1);
     });
 
     test('should match snapshot on loading', () => {
         const props = {...baseProps};
-        const wrapper = shallow(<TermsOfService {...props}/>);
+        const wrapper = shallow<TermsOfService>(<TermsOfService {...props}/>);
         wrapper.setState({loading: true});
         expect(wrapper).toMatchSnapshot();
     });
 
     test('should match snapshot, on accept terms', () => {
         const props = {...baseProps};
-        const wrapper = shallow(<TermsOfService {...props}/>);
+        const wrapper = shallow<TermsOfService>(<TermsOfService {...props}/>);
         wrapper.setState({loadingAgree: true});
         expect(wrapper).toMatchSnapshot();
     });
 
     test('should match snapshot, on reject terms', () => {
         const props = {...baseProps};
-        const wrapper = shallow(<TermsOfService {...props}/>);
+        const wrapper = shallow<TermsOfService>(<TermsOfService {...props}/>);
         wrapper.setState({loadingDisagree: true});
         expect(wrapper).toMatchSnapshot();
     });
 
     test('should call updateTermsOfServiceStatus on registerUserAction', async () => {
-        const wrapper = shallow(<TermsOfService {...baseProps}/>);
-        await (wrapper.instance() as TermsOfService).registerUserAction(true, jest.fn());
+        const wrapper = shallow<TermsOfService>(<TermsOfService {...baseProps}/>);
+        await wrapper.instance().registerUserAction(true, jest.fn());
         expect(baseProps.actions.updateMyTermsOfServiceStatus).toHaveBeenCalledTimes(1);
     });
 
     test('should match state and call updateTermsOfServiceStatus on handleAcceptTerms', () => {
-        const wrapper = shallow(<TermsOfService {...baseProps}/>);
-        (wrapper.instance() as TermsOfService).handleAcceptTerms();
+        const wrapper = shallow<TermsOfService>(<TermsOfService {...baseProps}/>);
+        wrapper.instance().handleAcceptTerms();
         expect(wrapper.state('loadingAgree')).toEqual(true);
         expect(wrapper.state('serverError')).toEqual(null);
         expect(baseProps.actions.updateMyTermsOfServiceStatus).toHaveBeenCalledTimes(1);
     });
 
     test('should match state and call updateTermsOfServiceStatus on handleRejectTerms', () => {
-        const wrapper = shallow(<TermsOfService {...baseProps}/>);
-        (wrapper.instance() as TermsOfService).handleRejectTerms();
+        const wrapper = shallow<TermsOfService>(<TermsOfService {...baseProps}/>);
+        wrapper.instance().handleRejectTerms();
         expect(wrapper.state('loadingDisagree')).toEqual(true);
         expect(wrapper.state('serverError')).toEqual(null);
         expect(baseProps.actions.updateMyTermsOfServiceStatus).toHaveBeenCalledTimes(1);
     });
 
     test('should call emitUserLoggedOutEvent on handleLogoutClick', () => {
-        const wrapper = shallow(<TermsOfService {...baseProps}/>);
-        (wrapper.instance() as TermsOfService).handleLogoutClick({preventDefault: jest.fn()} as unknown as React.MouseEvent<HTMLAnchorElement, MouseEvent>);
+        const wrapper = shallow<TermsOfService>(<TermsOfService {...baseProps}/>);
+        wrapper.instance().handleLogoutClick({preventDefault: jest.fn()} as unknown as React.MouseEvent<HTMLAnchorElement, MouseEvent>);
         expect(emitUserLoggedOutEvent).toHaveBeenCalledTimes(1);
         expect(emitUserLoggedOutEvent).toHaveBeenCalledWith('/login');
     });
